Fix Incorrect Currency in Mongoid Money Types
Workarea's default values for the Money fields in `Pricing::Override`
didn't previously change currency when `Money.default_currency` is
re-configured in process (like in the case of a multi-site application
with multiple currencies). Ensure that the correct currency is used by
using an Integer type as the default, which will get converted into a
Money type at runtime.

ECOMMERCE-7067