Customize Search Queries That Return an Exact Match (#22)
It's currently possible to customize search queries that return an exact
match, but instead of seeing the customized results when you run the
query, you'll be redirected to the product page since the
`StorefrontSearch::ExactMatches` middleware stops further middleware
execution and sets a redirect to the product path. To resolve the issue,
Workarea will now ignore this middleware if a customization is present
on the search response.

Discovered by @ryaan-anthony of **Syatt Media**. Thanks Ryan!