Fix Product URL In Breadcrumbs
The `storefront_url_for` method doesn't handle models other than taxons,
but the Schema.org helpers use it to render breadcrumb URLs in the
`BreadcrumbList` for any model that's in the breadcrumbs. To prevent
incorrect URLs from showing up in the breadcrumbs, the
`Navigation::Breadcrumbs` class has been modified to accept a model
object as its `:last` parameter, instead of just a name, to be added to
an arbitrary `Navigation::Taxon` created for the purpose of rendering
both the name and URL of the final navigation taxon. This wasn't needed
prior to the introduction of Schema.org's `BreadcrumbList`, because the
final URL of breadcrumbs was always left out. The helper methods that
render the breadcrumbs will continue to leave out the final taxon's URL,
but for breadcrumbs in Schema.org, the URL will now be included.

(#83)