Pass Options To `Storefront::UserActivityViewModel`
This was an oversight that got caught and fixed in the `flow-io` plugin,
but should really be in base since it will allow more control over the
product summaries on the recent views action. The `view_model_options`
were not getting passed into the `UserActivityViewModel`, and thus the
`ProductViewModel` instances that it creates, causing some stale content
to appear in the view.