[DOCS] Improve table of contents for docs
The table of contents that appears within each doc contains a link to
every h2-h6 in the document. Therefore, in longer docs, the TOC gets
quite crowded and stops communicating the overall structure of the doc.

Help readers maintain context by simplifying the TOC, limiting links to
h2-h3.

I spot-checked docs of various lengths and found this version of the TOC
more useful in all cases.

WORKAREA-96