Inject a hidden input for Select2 instances that allowClear
`WORKAREA.remoteSelects` works in a similar fashion where, to clear a
multiple-select Select2 instance we inject a hidden input before the
select menu in the DOM to be able to pass empty data along to the server
after the selection is made.

This is a little cleaner than Select2's default behavior, which forces
you to create a blank "placeholder" element, the value of which cannot
be an empty string. Weird, I know.

No changelog

WORKAREA-119