Add Guide for Installing Plugins
This guide explains how to install a plugin in the Workarea platform.
Although there may be some overlap with how this works in the rest of
the Rails community, there is no mention of how to connect to our
private gem server, which is not something most Rails developers know
how to do. It addresses the lack of mentions for the
`$BUNDLE_GEMS__WEBLINC__COM` environment variable in our documentation,
which is a knowledge gap for newcomers to Workarea and/or using private
gem servers.

WORKAREA-204