Force autoloading of BulkIndexProducts
app/workers/workarea/bulk_index_products.rb isn't getting autoloaded by
Rails. This causes a NameError to be raised for admin actions like
updating a product in the development environment.

This quick and dirty hack should be tested to see if it can be removed
after the update to Zeitwerk.