Fix locale not passed through in return redirect when not in URL
If a return_to parameter is generated without the locale, and a request
includes a parameter to switch locale, the locale is dropped causing the
request to revert to the default locale.

The original observed bug is switching locale in content editing and
seeing the request to save always redirect to the default locale.