Make Workarea compatible with Zeitwerk
This updates to the latest version of the rails-decorators gem, and
makes a few minor changes to allow Zeitwerk loading to function
correctly.

To check that your project is good to go for Zeitwerk, we recommend
running `rails zeitwerk:check` which will test your application to
determine if it's ready.