Add paranoid fallback for segment metrics lookup
Although this should never happen, giving a user incorrect segments
could have important consequences. If the email cookie is removed or
missing for some other reason, it doesn't hurt to fallback to looking up
based on the user model (even though this is an additional query) when
we know they're logged in.