Fix query caching in Releasable
When reloading a model to get an instance for a release, if the model
had already been loaded, a cached version of the model was returned.
This causes incorrect values on the instance you thought you were getting
for a release.

This first manifested as a bug where adding a featured product that
had a release change to make it active caused reindexing to make it
active but it shouldn't have been.