Fix Segments Workflow Setup Duplication
The setup form for the new custom segment workflow did not include the
ID of an existing segment (if persisted) in the form when submitted,
causing multiple duplicate segment records to be created when users go
back to the setup step in the workflow. None of the other steps are
affected because the ID appears in the URL, but the setup step does a
direct POST to `/admin/create_segments`, thus causing this problem.

WORKAREA-219