Setup PlaceOrderIntegrationTest in a Method
Currently, decorating the PlaceOrderIntegrationTest to edit the way its
set up (such as when adding an additional step) is impossible, you have
to basically copy everything out of the `setup` block and duplicate it
in your tests. Setup blocks should be methods anyway, so convert this to
a method and allow it to be decorated in host apps.