Fix Admin Configuration for Email Addresses
The hard-coded `config.email_from` and `config.email_to` settings
conflict with the out-of-box administrable configuration for the "Email
From" and "Email To" settings. This causes a warning for admins that
explain why the settings on "Email To" and "Email From" won't take
effect. Since the whole purpose of moving these settings to admin
configuration was to let admins actually change them, the
`config.email_from` and `config.email_to` settings have been removed
from both default configuration and the `workarea:install` generator.

WORKAREA-270