Set Default Inventory Policy to "Standard" in Create Product Workflow
When creating a new product through the workflow, setting the
"Inventory" on a particular SKU would still cause the `Inventory::Sku`
to be created with the "Ignore" policy rather than "Standard". Setting
inventory on a SKU now automatically causes the `Inventory::Sku` record
to be created with a policy of "Standard" so as to deduct the given
inventory to the varaint. When no inventory is given, Workarea will fall
back to the default of "Ignore".

WORKAREA-265