Fix Precision of Tax Rates UI
The `:step` values of the new/edit forms and precision configuration for
`#number_to_percentage` were not only rounding certain tax rates to an
incorrect number, but were also showing a bunch of insignificant zeroes
in the admin for tax rates. To resolve this, configure
`#number_to_percentage` to have 3-decimal precision, and strip all
insignificant zeroes from the display, leaving the admin with a much
nicer percentage display than what was presented before.

WORKAREA-278