Remove unnecessary Capybara blocking when testing content is not present
Capybara's `page.has_content?` and similar methods block until a timeout
is reached if they can't find the content. This is not what we want if
we're checking that the content does *not* exist, switch to using
`refute_text` in these scenarios.

The timeout is equal to the number of installed plugins and we have
client apps with 30+, which means that the 38 instances removed in this
commit could represent twenty minutes of unnecessary blocking in some
scenarios.

Thanks to James Anaipakos in
https://discourse.workarea.com/t/capybara-refute-assertions-waiting-full-default-wait-time/1610
for alerting me to the issue.