Check if a releasable has a localized active field before redefining it
If Workarea.config.localized_active_field is set to false, the active
field is redefined for each Releasable model to ensure the configuration
is honored. With inherited models like discounts, this can cause the
redefintion of active multiple times causing it to override custom active
behaviors for segments. Only redefining the method if its currently in
the models localized_fields list should ensure this does not happen.

WORKAREA-309