Improve password performance when running tests
Lowering the bcrypt cost lowers the time required to encrypt a password, at the
cost of increasing the speed at which an attacker can try to crack the password.
This is an acceptable tradeoff for improving performance of running tests. This
shaves about 5 minutes off of the admin tests.

Hat tip to Jeff Yucis for discovery.